+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This file contains highlights of what changes on each version of the [Free Mobil
 ## Version 3.0.0
 - Breaking change: raised the required [Dart](https://www.dartlang.org) version.
 - Breaking change: the class properties are now read-only.
+- Breaking change: removed the `toJson()` and `toString()` methods.
 - Changed licensing for the [MIT License](https://opensource.org/licenses/MIT).
 - Updated the package dependencies.